default logger is noop logger
